server/recommender/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
httpx
pydantic